requirements.txt
certifi            2025.10.5
charset-normalizer 3.4.4
colorama           0.4.6
contourpy          1.3.3
cycler             0.12.1
filelock           3.20.0
fonttools          4.60.1
fsspec             2025.10.0
idna               3.11
ImageIO            2.37.2
Jinja2             3.1.6
kiwisolver         1.4.9
lazy_loader        0.4
MarkupSafe         3.0.3
matplotlib         3.10.7
mpmath             1.3.0
networkx           3.5
numpy              2.3.4
opencv-python      4.11.0.86 (Xử lý ảnh và video)
packaging          25.0
pandas             2.3.3
pika               1.3.2 (Tương tác với RabbitMQ, 1.3.2 tương thích với tất cả phiên bản RabbitMQ)
pillow             12.0.0
pip                25.3
psutil             7.1.3
py-cpuinfo         9.0.0
pyparsing          3.2.5
python-dateutil    2.9.0.post0
pytz               2025.2
PyYAML             6.0.2 ( Đọc hoặc ghi file YAML)
requests           2.32.5 (Gửi HTTP requests, 2.32.3)
scikit-image       0.25.2
scipy              1.16.3
seaborn            0.13.2
setuptools         80.9.0
six                1.17.0
sympy              1.14.0
tifffile           2025.10.16
torch              2.9.0
torchvision        0.24.0
tqdm               4.67.1
typing_extensions  4.15.0
tzdata             2025.2
ultralytics        8.3.122 (Framework train model YOLO)
ultralytics-thop   2.0.18
urllib3            2.5.0
